requirements.txt
aptly-api-client==0.3.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
Django==5.2.4
django-debug-toolbar==6.0.0
django-widget-tweaks==1.5.0
environs==14.3.0
gunicorn==23.0.0
idna==3.10
iso8601==2.1.0
marshmallow==4.0.1
nodeenv==1.9.1
numpy==2.3.2
packaging==25.0
pyarrow==21.0.0
pylance==0.32.1
pyright==1.1.403
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
